Improve ActionRenderer2D
Attempted to provide more clarity in this refactor, so much happened but generally it's been simplified without any major noticeable performance impact or improvement.

Before I tear into this and get textures working, I really wanted to get some understanding and flush out a lot of broken or unused or unimportant concepts from this renderer, and I hope that this is generally an improvement.

The main clarity really comes in how the triangle processing pipeline works, and we added a small utility class to keep the matrix math out of the renderer, but yeah other than that too much to list..