Fix weird bug about input just pressed being checked twice returning false
idk what the deal is but like this one time I was checking up/down just pressed and it doesn't work if I check it twice? I've never been careful or mindful of checking just pressed more than once so idk if it's a bug or what the heck was happening but apparently we're ok now and the magic menu works.