Create ObjectShader
The whole shader set concept was something that has stuck around since our first journey into gpu 3d.
In reality, it was a complex mess of interweaving classes that eventually would just be a container system.
In the new setup, we have shaders with their own variants which are defined within their own classes.
ProgramManager can still cycle the visuals by configuring the shaders to use these variants.
It was crazy to come full circle, and finally tie up these loose ends, like, object shader is such a more digestable concept, and is more in line with the other shaders.