Overhaul SceneMode
SceneMode is now a bootstrap for SceneManager
SceneManager is the orchestrator of GameScenes
BaseGameScene encompasses the concepts of any scene in the game which is not classified as it's own mode.
Meaning Towns, traversal scenes, cutscenes, indoor areas, just general game areas which are not World mode.
In general, the GameScene itself is in control of the camera, and each of our scenes has different camera operation.
GameScenes are meant to allow for freedom of game location creation which is outside of the world mode scope.
At some basic level, the GameScenes have a camera, a character, physics world, and 3d level geometry.
GameScenes are meant to be flexible, but something like FishingMode is a concept which wouldn't be handled in this way.
We have a town, and indoor scenes to go into and exit from, a world portal to return us to the world from the town scene.
Random battle encounters are enabled for TownScene and disabled for HouseScene.
In it's current state, each GameScene has it's own corresponding SceneLayout which defines level geometry.
The TriggerPortal class is used to transition us to and from scenes.
It probably needs a look over to ensure everything is implemented in an agreeable way, but it is working currently.